requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-simple-history==3.4.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Pillow==10.1.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
